v1/requirements.txt
Flask>=0.12
python-smbus

